Implemented the possibleMoves method. While using the helper methods I saw that using more than 26 letters would create a huge problem with the indexing scheme. For this version I dropped the possibility to use a spreadsheet like naming (i.e "AA", "AB", etc.)
--- ChessboardSplitColumnRowTest.java ---
Updated the test cases to make "aa1" and "aa11" invalid values

--- ChessboardNameToColumnTest.java ---
Udpated the test cases to make "aA", "AA", "Abc", and "AbC" invalid values
Added a test method that uses the value and the expected result from the parameter in the reverse order to test tha value to name conversion

--- ChessboardPossibleMovesTest.java ---
Added more test cases for the knight movements
Sorted the resulting List of positions so that they can be compared

--- Chessboard.java ---
Created constants to keep track of the board bounds
Added two attributes that will hold the current board size, currently initialized to the minimum size (8x8)
Updated the splitColumnRow to consider the column with a single char and use a parseInt to check if the row is in correct notation
Updated nameToColumn to handle a single char
Added the columnToName that reverts from the int value to the name of the position
Implemented the possibleMoves with the hardcoded moves of the knight, this method calls the other utilitary methods. The row index is decremented to cope with zero indexing. From the column and row we create a Position object and calculate the eight possible moves of the knight. After that we check which positions are valid regarding the current board and convert the Positions to algebraic notation to add to the resulting List of String.
Added the isValid method to check if a Position is within the board bounds
Added the positionToAlgebraic method that converts a Position to a String in the algebraic notation
Added the Postion inner class for organizing the calculations