Fixes #11 Using the interface for ChessPieceMover calculates the N-depth possible moves from a given start for the given piece
--- Position.java ---
Implesmented the equals and hashCode methods to allow for comparison when inside a List
Also improved the toString method

--- ChessboardPossibleMovesTest.java ---
Adapted the test to include the depth information
Created an extra test for depth 2

--- Chessboard.java ---
Removed the calculateValidMovesForPiece
Added the level control so that we can calculate all the possible moves based on the previous result. Level zero is considered with only the start Position. Follow the candidate List executing the moves method to get all the possible moves for each of the candidate positions. Test which positions are valid for this Chessboard and which are not in the level result list yet and add the final level result to the HashMap, in the end aggregate all the positions removing duplicates. This de-duplication has to be review it is not efficient as it could be.