Fixes #14 Added the initial version of the web client. It uses Spring Boot with JSF
--- pom.xml ---
Added the dependencies for Spring and JSF

--- faces-config.xml ---
Added the Spring ELResolver

--- web.xml ---
Added the Spring Listeners

--- Application.java ---
Spring application that loads the container for the Servlets

--- Cell.java ---
Represents a single board cell, added many attributes to solve the painting on the screen. Needs refactoring

--- Row.java ---
Represents a Row in the board filled with Cells

--- ChessboardBean.java ---
Bean for user interface controlling, creates the board in terms of a List of Rows (wich is a List of Cells). Sets the Cell attributes
Defined the methods to add/remove Rows/Columns
Added a method to reset the board
Added a method to call the Chessboard service and get a list o moves for the position that the user clicked

--- index.xhtml ---
Defined the table to present white and black buttons for the user to choose where to place the knight